hw4/311513058/311513058/requirements.txt
accelerate==0.18.0
asttokens==2.2.1
backcall==0.2.0
certifi==2019.11.28
chardet==3.0.4
cmake==3.26.1
coloredlogs==15.0.1
comm==0.1.3
cycler==0.10.0
dbus-python==1.2.16
debugpy==1.6.6
decorator==5.1.1
diffusers==0.12.1
distro==1.4.0
distro-info===0.23ubuntu1
executing==1.2.0
filelock==3.10.3
flatbuffers==23.3.3
huggingface-hub==0.13.3
humanfriendly==10.0
idna==2.8
importlib-metadata==6.1.0
invisible-watermark==0.1.5
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
Jinja2==3.1.2
jupyter-client==8.1.0
jupyter-core==5.3.0
kiwisolver==1.0.1
lit==16.0.0
MarkupSafe==2.1.2
matplotlib==3.1.2
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.0
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
olefile==0.46
onnx==1.13.1
onnxruntime==1.14.1
opencv-python==4.7.0.72
packaging==23.0
pandas==2.0.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
PyGObject==3.36.0
pyparsing==2.4.6
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
regex==2023.3.23
requests==2.22.0
requests-unixsocket==0.2.0
scipy==1.10.1
six==1.14.0
ssh-import-id==5.10
stack-data==0.6.2
sympy==1.11.1
tokenizers==0.12.1
torch==2.0.0
torchaudio==2.0.1
torchdata==0.6.0
torchtext==0.15.1
torchvision==0.15.1
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.19.2
triton==2.0.0
typing-extensions==4.5.0
tzdata==2023.3
unattended-upgrades==0.1
unet==0.7.7
urllib3==1.25.8
wcwidth==0.2.6
zipp==3.15.0
